+++ src/routes/(admin)/admin/eventsadmin/create/+page.svelte
@@ -1,16 +1,12 @@
 <script lang="ts">
-	import { goto } from '$app/navigation';
+	import { beforeNavigate, goto } from '$app/navigation';
 	import { eventsColRef } from '$lib/firebase/firebaseConfig';
 	import { addDoc } from 'firebase/firestore';
 
 	//import { notificationStore } from '$lib/stores/notifications';
 	import { EditMode, EditModeStore, type DomainEvent } from '$lib/stores/ObjectStore';
-	import {
-		selectedImage,
-		imageExists,
-		existingImageUrl,
-		resetImageselection,
-	} from '$lib/stores/ImageSelectionStore';
+	import { selectedImage, imageExists, existingImageUrl, resetImageselection } from '$lib/stores/ImageSelectionStore';
+	import { Messages } from '$lib/utils/messages';
 
 	import { validateEventData, buildTimeStamp } from '$lib/services/validateForm';
 	import { eventFormService } from '$lib/services/EventFormService';
@@ -18,6 +14,9 @@
 	import ToastContainer from '$lib/components/ToastContainer.svelte';
 	import EventForm from '$lib/components/EventForm.svelte';
 	import { notificationStore } from '$lib/stores/notifications';
+	import { resetSelectedLocation } from '$lib/stores/LocationsStore';
+
+	let pageHasUnsavedChanges = $state(false);
 
 	const saveNewEvent = async (newEvent: DomainEvent) => {
 		if (validateEventData(newEvent)) {
@@ -33,18 +32,57 @@
 		try {
 			await addDoc(eventsColRef, thisEvent);
 			EditModeStore.set(EditMode.Empty);
-			notificationStore.addToast('success', 'Event created successfully!');
+			notificationStore.addToast('success', Messages.SAVESUCCESS);
 			resetImageselection();
 			goto('/admin/eventsadmin');
 		} catch (error) {
-			notificationStore.addToast('error', 'Failed to create event. Please try again.');
+			notificationStore.addToast('error', Messages.SAVERROR);
 			console.error('Error creating event: ', error);
 		}
 	};
+
+	const saveDraft = async (thisEvent: DomainEvent) => {
+		if (!thisEvent) return;
+		try {
+			await addDoc(eventsColRef, thisEvent);
+			EditModeStore.set(EditMode.Empty);
+			notificationStore.addToast('success', Messages.DRAFTSUCCESS);
+			resetImageselection();
+			goto('/admin/eventsadmin');
+		} catch (error) {
+			notificationStore.addToast('error', Messages.DRAFTERROR);
+			console.error('Error creating event: ', error);
+		}
+	};
+
+	const handleCancel = () => {
+		EditModeStore.set(EditMode.Empty);
+		resetSelectedLocation();
+		goto('/admin/eventsadmin');
+	};
+
+	const handleUnsavedChangesUpdate = (formHasUnsavedChanges: boolean): void => {
+		pageHasUnsavedChanges = formHasUnsavedChanges;
+	};
+	/**
+	 * TODO: replcae dialog with the global modal (to be delivered)
+	 */
+	beforeNavigate(({ cancel }) => {
+		if (pageHasUnsavedChanges) {
+			if (!confirm('You have unsaved changes. Are you sure you want to leave this page?')) {
+				cancel();
+			}
+		}
+	});
 </script>
 
 <div>
-	<EventForm onCreateNew={saveNewEvent} />
+	<EventForm
+		onCreateNew={saveNewEvent}
+		onCancel={handleCancel}
+		onSaveDraft={saveDraft}
+		onUnsavedChangesUpdate={handleUnsavedChangesUpdate}
+	/>
 </div>
 
 <ToastContainer />