+++ src/routes/(admin)/admin/newsadmin/[newsId]/+page.svelte
@@ -1,15 +1,18 @@
 <script lang="ts">
-
+	import { beforeNavigate, goto } from '$app/navigation';
 	import { updateDoc, type DocumentReference, type DocumentData } from 'firebase/firestore';
 
 	import { type News, EditModeStore } from '$lib/stores/ObjectStore';
 	import { selectedImage, imageExists, existingImageUrl } from '$lib/stores/ImageSelectionStore';
-	
+
 	import { newsFormService } from '$lib/services/NewsFormService';
 	import { notificationStore } from '$lib/stores/notifications';
+	import { Messages } from '$lib/utils/messages';
 
 	import ToastContainer from '$lib/components/ToastContainer.svelte';
 	import NewsForm from '$lib/components/NewsForm.svelte';
+
+	let pageHasUnsavedChanges = $state(false);
 
 	interface Props {
 		data: {
@@ -20,7 +23,24 @@
 
 	let { data }: Props = $props();
 
+	const handleCancel = () => {
+		EditModeStore.set('');
+		goto('/admin/newsadmin');
+	};
 
+	const handleUnsavedChangesUpdate = (formHasUnsavedChanges: boolean): void => {
+		pageHasUnsavedChanges = formHasUnsavedChanges;
+	};
+	/**
+	 * TODO: replcae dialog with the global modal (to be delivered)
+	 */
+	beforeNavigate(({ cancel }) => {
+		if (pageHasUnsavedChanges) {
+			if (!confirm('You have unsaved changes. Are you sure you want to leave this page?')) {
+				cancel();
+			}
+		}
+	});
 
 	const updateNews = async (updatedNews: News) => {
 		try {
@@ -38,16 +58,39 @@
 			const newsData = { ...thisNews } as DocumentData;
 			await updateDoc(data.docRef, newsData);
 			EditModeStore.set('');
-			notificationStore.addToast('success', 'News article updated successfully!');
+			notificationStore.addToast('success', Messages.UPDATESUCCESS);
+			goto('/admin/newsadmin');
 		} catch (error) {
-			notificationStore.addToast('error', 'Failed to update news article. Please try again.');
+			notificationStore.addToast('error', Messages.UPDATEERROR);
+			console.error('Error updating the news: ', error);
+		}
+	};
+
+	const handleSaveDraft = async (thisNews: News) => {
+		try {
+			if (!data.docRef) {
+				throw new Error('No document reference provided');
+			}
+			const newsData = { ...thisNews } as DocumentData;
+			await updateDoc(data.docRef, newsData);
+			EditModeStore.set('');
+			notificationStore.addToast('success', Messages.UPDATESUCCESS);
+			goto('/admin/newsadmin');
+		} catch (error) {
+			notificationStore.addToast('error', Messages.UPDATEERROR);
 			console.error('Error updating the news: ', error);
 		}
 	};
 </script>
 
 <div>
-	<NewsForm thisNews={data.newsItem} onUpdate={updateNews} />
+	<NewsForm
+		thisNews={data.newsItem}
+		onUpdate={updateNews}
+		onSaveDraft={handleSaveDraft}
+		onCancel={handleCancel}
+		onUnsavedChangesUpdate={handleUnsavedChangesUpdate}
+	/>
 </div>
 
 <ToastContainer />